fix(event-bus): inline noopWorker into main.go for single-file build convention
The repo's build convention is 'go build cmd/crm-api/main.go' (single-file),
used by the Makefile (6 sites), CI ARM64 cross-compile, and the Playwright
E2E webServer. Splitting noopJobArgs/noopWorker into a sibling river.go
caused 'undefined: noopWorker' in both the ARM64 build and E2E webServer
boot on PR #279.

Move the types back inline in main.go (as the original plan intended) and
delete river.go. Added a comment documenting why they stay inline.

Fixes CI failures on #279.
